requirements.txt
flask
flask_sqlalchemy
flask_bootstrap
pymupdf
openai
markdown-it-py
mdit-py-plugins
weasyprint
python-dotenv
gunicorn
